backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
httpx==0.27.0
supabase==2.4.6
spacy==3.7.4
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1